Create bitwarden-core with macro and uniffi
