+++ crates/bitwarden-core/Cargo.toml
@@ -1,0 +1,26 @@
+[package]
+name = "bitwarden-core"
+description = """
+Internal crate for the bitwarden crate. Do not use.
+"""
+
+version.workspace = true
+authors.workspace = true
+edition.workspace = true
+rust-version.workspace = true
+homepage.workspace = true
+repository.workspace = true
+license-file.workspace = true
+keywords.workspace = true
+
+[features]
+uniffi = ["dep:uniffi"]
+
+[dependencies]
+chrono = { version = ">=0.4.26, <0.5", default-features = false }
+uniffi = { version = "=0.27.2", optional = true }
+uuid = { version = ">=1.3.3, <2.0", features = ["serde"] }
+thiserror = ">=1.0.40, <2.0"
+
+[lints]
+workspace = true
