+++ crates/bitwarden-core/src/error.rs
@@ -1,0 +1,19 @@
+use thiserror::Error;
+
+#[derive(Debug, Error)]
+#[error("The response received was missing a required field: {0}")]
+pub struct MissingFieldError(pub &'static str);
+
+/// This macro is used to require that a value is present or return an error otherwise.
+/// It is equivalent to using `val.ok_or(Error::MissingFields)?`, but easier to use and
+/// with a more descriptive error message.
+/// Note that this macro will return early from the function if the value is not present.
+#[macro_export]
+macro_rules! require {
+    ($val:expr) => {
+        match $val {
+            Some(val) => val,
+            None => return Err($crate::MissingFieldError(stringify!($val)).into()),
+        }
+    };
+}
