+++ crates/bitwarden-core/src/lib.rs
@@ -1,0 +1,7 @@
+#[cfg(feature = "uniffi")]
+uniffi::setup_scaffolding!();
+#[cfg(feature = "uniffi")]
+mod uniffi_support;
+
+mod error;
+pub use error::MissingFieldError;
