+++ crates/bitwarden-core/src/uniffi_support.rs
@@ -1,0 +1,32 @@
+use uuid::Uuid;
+
+use crate::UniffiCustomTypeConverter;
+
+type DateTime = chrono::DateTime<chrono::Utc>;
+uniffi::custom_type!(DateTime, std::time::SystemTime);
+
+impl UniffiCustomTypeConverter for chrono::DateTime<chrono::Utc> {
+    type Builtin = std::time::SystemTime;
+
+    fn into_custom(val: Self::Builtin) -> uniffi::Result<Self> {
+        Ok(Self::from(val))
+    }
+
+    fn from_custom(obj: Self) -> Self::Builtin {
+        obj.into()
+    }
+}
+
+uniffi::custom_type!(Uuid, String);
+
+impl UniffiCustomTypeConverter for Uuid {
+    type Builtin = String;
+
+    fn into_custom(val: Self::Builtin) -> uniffi::Result<Self> {
+        Uuid::parse_str(val.as_str()).map_err(|e| e.into())
+    }
+
+    fn from_custom(obj: Self) -> Self::Builtin {
+        obj.to_string()
+    }
+}
