+++ crates/bitwarden-core/uniffi.toml
@@ -1,9 +1,9 @@
 [bindings.kotlin]
-package_name = "com.bitwarden.bitwarden"
+package_name = "com.bitwarden.core"
 generate_immutable_records = true
 android = true
 
 [bindings.swift]
-ffi_module_name = "BitwardenBitwardenFFI"
-module_name = "BitwardenBitwarden"
+ffi_module_name = "BitwardenCoreFFI"
+module_name = "BitwardenCore"
 generate_immutable_records = true
