+++ crates/bitwarden-crypto/Cargo.toml
@@ -18,7 +18,6 @@ default = []
 
 uniffi = ["dep:uniffi"]  # Uniffi bindings
 no-memory-hardening = [] # Disable memory hardening features
-test = []                # Test methods
 
 [dependencies]
 aes = { version = ">=0.8.2, <0.9", features = ["zeroize"] }
