+++ crates/bitwarden-exporters/Cargo.toml
@@ -28,8 +28,5 @@ serde_json = ">=1.0.96, <2.0"
 thiserror = ">=1.0.40, <2.0"
 uuid = { version = ">=1.3.3, <2.0", features = ["serde", "v4"] }
 
-[dev-dependencies]
-bitwarden-crypto = { workspace = true, features = ["test"] }
-
 [lints]
 workspace = true
