+++ crates/bitwarden/src/admin_console/policy.rs
@@ -1,12 +1,13 @@
 use std::collections::HashMap;
 
 use bitwarden_api_api::models::PolicyResponseModel;
+use bitwarden_core::require;
 use schemars::JsonSchema;
 use serde::{Deserialize, Serialize};
 use serde_repr::{Deserialize_repr, Serialize_repr};
 use uuid::Uuid;
 
-use crate::error::{require, Error, Result};
+use crate::error::{Error, Result};
 
 #[derive(Serialize, Deserialize, Debug, JsonSchema)]
 pub struct Policy {
