+++ crates/bitwarden/src/auth/login/access_token.rs
@@ -1,6 +1,7 @@
 use std::path::{Path, PathBuf};
 
 use base64::{engine::general_purpose::STANDARD, Engine};
+use bitwarden_core::require;
 use bitwarden_crypto::{EncString, KeyDecryptable, SymmetricCryptoKey};
 use chrono::Utc;
 use schemars::JsonSchema;
@@ -14,7 +15,7 @@ use crate::{
         AccessToken, JWTToken,
     },
     client::{LoginMethod, ServiceAccountLoginMethod},
-    error::{require, Error, Result},
+    error::{Error, Result},
     secrets_manager::state::{self, ClientState},
     Client,
 };
