+++ crates/bitwarden/src/error.rs
@@ -15,6 +15,8 @@ use thiserror::Error;
 
 #[derive(Debug, Error)]
 pub enum Error {
+    #[error(transparent)]
+    MissingFieldError(#[from] bitwarden_core::MissingFieldError),
     #[error("The client is not authenticated or the session has expired")]
     NotAuthenticated,
 
@@ -26,8 +28,6 @@ pub enum Error {
 
     #[error("The response received was invalid and could not be processed")]
     InvalidResponse,
-    #[error("The response received was missing some of the required fields: {0}")]
-    MissingFields(&'static str),
 
     #[error("Cryptography error, {0}")]
     Crypto(#[from] bitwarden_crypto::CryptoError),
@@ -153,19 +153,5 @@ macro_rules! impl_bitwarden_error {
 }
 impl_bitwarden_error!(ApiError);
 impl_bitwarden_error!(IdentityError);
-
-/// This macro is used to require that a value is present or return an error otherwise.
-/// It is equivalent to using `val.ok_or(Error::MissingFields)?`, but easier to use and
-/// with a more descriptive error message.
-/// Note that this macro will return early from the function if the value is not present.
-macro_rules! require {
-    ($val:expr) => {
-        match $val {
-            Some(val) => val,
-            None => return Err($crate::error::Error::MissingFields(stringify!($val))),
-        }
-    };
-}
-pub(crate) use require;
 
 pub type Result<T, E = Error> = std::result::Result<T, E>;
