+++ crates/bitwarden/src/uniffi_support.rs
@@ -3,7 +3,9 @@ use std::num::NonZeroU32;
 use bitwarden_crypto::{AsymmetricEncString, EncString};
 use uuid::Uuid;
 
-use crate::UniffiCustomTypeConverter;
+type DateTime = chrono::DateTime<chrono::Utc>;
+uniffi::ffi_converter_forward!(DateTime, bitwarden_core::UniFfiTag, crate::UniFfiTag);
+uniffi::ffi_converter_forward!(Uuid, bitwarden_core::UniFfiTag, crate::UniFfiTag);
 
 uniffi::ffi_converter_forward!(NonZeroU32, bitwarden_crypto::UniFfiTag, crate::UniFfiTag);
 uniffi::ffi_converter_forward!(EncString, bitwarden_crypto::UniFfiTag, crate::UniFfiTag);
@@ -12,32 +14,3 @@ uniffi::ffi_converter_forward!(
     bitwarden_crypto::UniFfiTag,
     crate::UniFfiTag
 );
-
-type DateTime = chrono::DateTime<chrono::Utc>;
-uniffi::custom_type!(DateTime, std::time::SystemTime);
-
-impl UniffiCustomTypeConverter for chrono::DateTime<chrono::Utc> {
-    type Builtin = std::time::SystemTime;
-
-    fn into_custom(val: Self::Builtin) -> uniffi::Result<Self> {
-        Ok(Self::from(val))
-    }
-
-    fn from_custom(obj: Self) -> Self::Builtin {
-        obj.into()
-    }
-}
-
-uniffi::custom_type!(Uuid, String);
-
-impl UniffiCustomTypeConverter for Uuid {
-    type Builtin = String;
-
-    fn into_custom(val: Self::Builtin) -> uniffi::Result<Self> {
-        Uuid::parse_str(val.as_str()).map_err(|e| e.into())
-    }
-
-    fn from_custom(obj: Self) -> Self::Builtin {
-        obj.to_string()
-    }
-}
